Abandon the separate submodule idea
